server/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
torch>=2.6
transformers>=4.51
accelerate>=1.6
umap-learn>=0.5
scikit-learn>=1.6
numpy>=1.26
pydantic>=2.10
